Moves OpportunityAppSchema into opportunity_model.py This allows you to avoid a circular import error while enabling two-way nesting between opportunity and opportunity_app
